requirements.txt
fastapi
uvicorn[standard]
moderngl
imageio[pyav]
pillow
python-multipart
